Move TreeNodeType to its own package
The types of Tree Nodes is too core of a concept, and used in too many
locations to be allowed to persist in the ast package. It prevents ast
from importing many useful packages - such as veterrors - since it would
cause a dependency loop.

The actual unit of use which most importers care about is TreeNodeType.

This CL separates TreeNodeType into its own package, node, and renames
it "Type" to avoid path stuttering.

Change-Id: Id741cd1a0e52be148cead43cf15252b1f2aa7ad7