Import the CNRM IAM test and scaffolding
This change imports the IAM test that was written for CNRM, alongside
the CNRM testing scaffolding it uses. This scaffolding is based on
Kubebuilder and so does not have an external cluster dependency, but did
require a GCP dependency so the test can go and check the IAM policy was
created as intended.

Please _do not_ review the CNRM scaffolding in
pkg/bespin-controllers/test. This should be considered vendored code.

The CNRM test is currently in review by CNRM team, and comments from
that CL will be merged into this one where possible: some changes were
required to integrate with the resources as written, which we made
slightly different for CNRM. CL at: https://cnrm-review.git.corp.google.com/c/cnrm/+/3480

Fixes: http://b/123090656
Change-Id: Id8befc8d62ebc9fcc22fc05d4c3b915b1e219f94