Pin Terraform to v.0.11.11
This change pins Terraform to v.0.11.11, as there is a breakage of
Terraform at HEAD with the Google TF plugin. `go get` only pulls from
HEAD, and Nomos does not use Go modules, so the options were bring in a
large dependency into vendor/, or manually clone the repo. This takes
the second path.

Fixed: http://b/123282741
Change-Id: I93c125fb837ec7998c1f69a36e6a45b3e7330a6e