remove unneeded checks in e2e test setup
We only need to ensure that policynodes are based on the latest changes from the repo and that they don't have an error status. We already ensure the corresponding namespaces and resources have been synced by the time those fields are written.
I've seen a speedup of about 1 minute in e2e tests after this change.

Fixes: 123408511
Change-Id: I65a4e9f491595e5ccf8b17111da54e6869ae0b56