Add ActionOnError() to Visitor
Visitors don't have a way to indicate what should be done when they
produce errors. This means Visitors which are just performing validation
cause parsing to stop immediately when it is more useful to show all
errors.

This CL adds ActionOnError() to Visitors, which allows the Visitors to
signal what should be done when they produce errors. For now there are
two possible actions, StopOnError and ContinueOnError. More may be added
later.

Change-Id: Ideae795f505d56cc4d7d89799cb00c293d45bda1