Move tree construction to Visitor
While this CL is large, note that 700 of the 900 modified lines are test
code.

Tree construction has the potential to fail, but there isn't any logical
reason that it should be able to.

Also, tree construction is done as a special step which prevents
expanding the visitor model to steps done before creating the policy
hierarchy tree.

This CL makes a Visitor that handles creating the policy tree and
filling it with objects. We can now start moving namespaces/ validation
to Visitors.

This CL also adds much-needed test infrastructure for creating
hierarchical trees of objects. Wtih treetesting.BuildTree, all that is
required for specifying objects in the policy hierarchy is literally
listing the FileObjects under test. As a later CL, we may decide to add
handling system/, cluster/, and clusterregistry/.

As a side-effect, this CL makes us completely ignore directories with no
objects we care about. This is cleaner, but it is a functionality
change. No one should be depending on this, though.

Bug: 123767573
Bug: 123657136

Change-Id: If48b2de55471e713f1a29625b047039fbe8071f7