Move Namespace kind validation to a Visitor
There's a lot of logic specific to Nomos that is automatically used on
Bespin respositories.

This CL moves Namespace kind validation to a Visitor.

This CL removes specifying the Node type when adding it to a Tree,
requiring the caller to do this as a separate step when adding the node.
The default type is AbstractNamespace.

This significantly simplifies processNamespaces and paves the way to
building the hierarchy tree with a Visitor.

Change-Id: I496e25d03fa181901cb6afe51a2c964a0ad39c62