Keep Namespace and NamespaceSelector Resources
We currently immediately discard Namespace and NamespaceSelector
resources rather than putting them in the policy hierarchy. This means
we can't visit them with Visitors, as they are fully unavailable.

This CL removes the guard which prevents writing Namespace and
NamespaceSelector objects to TreeNode.Objects, allowing all Visitors
access to those types.

This CL also adds EphemeralResourceRemover, which is a Visitor that
removes these types which are only used client-side and not uploaded to
K8S.

Bug: 123657136
Change-Id: I022ecad8efd8119ace6dc3a052d53596e4900942