Finish moving namespaces/ validation to Visitors
This CL is purely refactoring namespaces/ validation to Visitors. Most
of this is purely mechanical and not interesting - mostly resulting in
lots of whitespace changes and the same refactoring over-and-over to
convert to Visitors.

The only interesting one is duplicate_name_validator. While moved to a
Visitor, it still does not support Selectors. That change is out of
scope of this CL.

Notably, since metadata validators were also used by other top-level
directories, that validation is now all in Visitors.

Changes to examples/errors/expected_errors.txt are expected because
there is now one error it catches before constructing the hierarchy
tree, and constructing the hierarchy tree requires no previous errors.

Change-Id: Idf5ecb876ca52cf6f6c0adf3386c0932d5d3c825