Set APIInfo in Root.Data
A lot of nomos-specific logic is unable to be converted into Visitors
because it is special-cased. We need to move away from steps that rely
on intermediate variables in Parser.parse(), and instead rely on
transforming the hierarchy tree as a way of caching what we need.

This CL adds setter/accessor methods which are the only way to retreive
this meta.APIInfo value. It also adds tests that verify adding scope
works as intended.

Change-Id: I3857a61773548e4c1f4424cde419ec6ff5ef522c