Configure gitcookies for prow job
Since we switched to the new prow test runner, we need to configure git
properly.

Testing situation:
This is tricky to test locally since it has to do with the way the prow
environment is set up and the permissions available therein. I will
coordinate with Sen to test whether this works, and if it does not, the
next step would be to start a staging job (or maybe use the canary?) to
debug further.

Change-Id: Ic7fc8d1bc3dfe81322030345c704d2ae4c8f89e9