Move name duplicate detection to a visitor
We've been doing validation of duplicate names in sevearal places, and
in some cases unnecessarily.

This CL runs name duplication validation on flattened policy nodes.

This CL forces many changes to tests in parser_test.go. Some of them are
expected - such as requiring Namespaces in order to cause the error. In
others I managed to get the test to work, but am not sure if that is the
intent.

Change-Id: I2be8ebe95eca7c29f215d567da785c5a20819ae8