Reject name collisions in Abstract Namespaces
We currently allow resources of the same Group/Kind to share the
same name in Abstract Namespaces so long as they are not materialized in
the same Namespace. This makes things like uniquely referencing resources
difficult. Also, it is basically using NamespaceSelectors for resource
customization, which we don't want to allow.

This CL verifies that no resources of the same Group/Kind share the same
metadata.name.

Change-Id: I7b5ae1d7ee6e1620a1bbc7a8626449849072cbf4