Add error for improper Nomos installation
The nomos CLI does not ouptut a human-readable error if Nomos is not
installed on the cluster in the current kubectl context.

This CL adds a sanity check to see if Nomos is installed, and a
human-readable/actionable message hinting at what to do.

If Nomos is currently being installed or uninstalled and the cluster is
in an inconsistent state, strange errors may still be printed. This
state lasts for at most several tens of seconds until the operator has
installed all required CRDs.

Bug: 120566481

Change-Id: I063ee23c555639c8431f8d757238f2283710931c