Improve e2e trap handling
Currently, killing a test in the middle successfully uninstalls but then
hangs forever. This change makes a tweak to trap a signal that will kill
the docker container (SIGINT 1). Also use our waiting function to wait
for nomos-system-test to come down, which means it won't wait forever
and makes use of our nice bash library.

Change-Id: I87d78be6b2f6233742dbfef037743fbe03ef9e3c