e2e: determines nomos is running using deployments
Old code used the state of the underlying pods, which is part but not
the entire state of a deployment.

I initially thought that it was the race between deployment availability
and uninstallation that caused b/124329060, but after implementing this
I found that the bug is still there.  I'd like to commit this
nevertheless, as it removes a confounding factor in the work on the bug.

Bug: 124329060
Change-Id: I0de0996e1800c92d83b921143f409c7414e3464b