Write test builders with functional options
It is getting to be a pain to write individual test methods to account
for every variant of AST possible.

This CL writes an interface for generating ASTs and FileObjects using a
functional interface.

This does duplicate code elsewhere, and after this CL is approved I'll
send individual refactoring CLs for each other test to move over to this
model.

This CL also refactors four example tests to show how it works. Note
that in general this CL reduces line length of test code by half.

Also note the convention that there is the nearly blank line, "),"
between tests. This is there to break up the tests visually so it's
clear where tests begin and end.

This CL does not break any existing tests.

Change-Id: I531655da30493e0cc9fb72cebecd5396471a3ccf