docs: more nomos-operator troubleshooting
I found that the old instructions were not quite enough
to unwedge a failed bless release process, so I tried
to rework with some more detailed information.

It's not verified, because `dep` failures seem nondeterministic
and at some point they just went away for me, but at least the menu
of options to try is a bit larger.

Change-Id: I3055a35def3068dd2be0aa7b85f7911b5507a141