Templatize usages of "nomos"
It's going to be a pain to rename every text instance of "nomos" and its
variants to something else. We need to reduce this future pain and allow
for people to not necessarily stick with what they pick.

This CL templatizes the usages of the string "nomos" in many places
throughout our codebase.

Many of these were one-offs where the string was just literally typed.

Bug: 120799315

Change-Id: I75c37d382b7871e5ceb81f8b21186b5effec75f5