Implement default file path generation
We have to have an output path for files in order to write resource
YAMLs to the hard drive.

This CL implements the logic that determines the unique file each
resource goes to.

Handling collections of resources which could not actually be on a
single cluster is out of scope of this logic.

Change-Id: Iabea89fed3be321fbc8901372fa3ee8d7b4dba64