Fix the issue that e2e tests combine stderr with stdout.
This combining is caused by pty, which is what docker uses to
implement --tty. We use --tty to fix the issue that docker doesn't
respect Ctrl+C. So this fixes that issue by adding tini to the
docker image.

tini is a minimal init. It does the things that PID 1 should do
(signal handling and zombie reaping) that bash isn't good at.

Bug: 126230640
Change-Id: Ia7f96903cb11accaf8cc0776d2aee987d8e8ed0f