Write ResourceLister
In order to list the resources in a Kubernetes cluster and write them to
the disk, we have to make the appropriate API calls and convert them to
a format our libraries work with.

This CL implements minimal interfaces that list resources from a
Kubernetes cluster so that tests, and potentially other implementers,
don't need to implement many methods they don't need.

Also implements DynamicResourcer, which has the single line of code
which won't be unit tested since it's just method forwarding  and
implementing the full interface would be a *lot* of code.

Bug: 126431492

Change-Id: I052235a9ed31005f0b5c5a741d36485f27a4fdbf