syncer: eliminate update conflicts
The old code did not cancel reconcile operations when the syncer closed
the "end" channel.  But, since at closing of the "end" channel all
informers stopped working, the reconcile operation would be re-reading a
stale object from cache at all times.

New code converts the end channel into a cancelable context and passes
that context into the reconcilers to replace the ambient context.  This
way, if the "end" channel is closed, the converted context gets canceled
and the reconcilers exit gracefully.

Bug: 77981474
Change-Id: Ia3aeedf9cafc1b34fc08423db8f3d99b6cd3b688