Reduce test code duplication in parser_test.go
Most tests aren't testing the behavior of the Repo resource, so they
shouldn't have to declare it. It's just noise that makes it harder to
understand what is being tested.

Ditto with all tests having to specify "foo" as root.

Ditto tests unnecessarily declaring default HierarchyConfig for every
resource when the tests aren't testing that.

Change-Id: Ia60fbe5b4059571e98aa7029e8bc585a28b3c463