Prune empty YAML lists and maps
As we delete fields and values from resources, this sometimes leaves
behind empty lists and maps. This leaves behind ugly nil maps and empty
lists that do not change the meaning of the resource file.

This CL implements Prune(), a Mutator that recursively removes empty
lists and maps from unstructured.Unstructured.

Bug: 128354078
Change-Id: Ie17f142a041002f125a40ca5e42888b8758ee45a