Define nomospath.Path
Often paths aren't actually relative to a Nomos Root.

This CL handles that case by defining a typesafe Path and using that in
ast.FileObject instead of nomospath.Relative. This allows representing
ast.FileObjects without misleadingly claiming that there is a literal
nomos Root the FileObject exists at.

The principle here is objects should only have, and be passing around,
the data that they actually use. The path relative to the OS root is
only used when reading files, and this feature is never used in the
context of an existing FileObject. Thus, FileObjects, only need a
typesafe Path.

Change-Id: Ifdc0a72ab61f9accd558925d368e0ee9359436f7