Makefile.release: fixes user label on a robot
Per gsutil docs, one must use `serviceAccount:` as the
schema name for the user URI if the user is a service
account.

Else you get the error like this:
BadRequestException: 400
user:425329972751-compute@developer.gserviceaccount.com was marked as a
user but is actually a serviceAccount

I'm sure there's a very good reason why this is so, but doesn't really
make for a swell user experience.  Oh, well.

Bug: 128694606
Change-Id: I4c4fc97532bd6b0b1b22d732ab1733e32515ef09