syncer: allow syncing the 'default' namespace
The old code disallowed managing the namespace 'default' as well as
'kube-*'.  This change removes the restriction from 'default', with the
intention that we slowly lift the restriction from other namespaces that
need such treatment as well.

In the process, added some logging and test changes that should make the
testing and debugging more robust.

Added integration tests that confirm the new functionality, and confirms
that unmanaged resources in default are not affected. Old code would not
have allowed us to touch the 'default' namespace.

A change proposal here:
https://docs.google.com/document/d/1I6PqEJGBMjC2xAgKMsKPUVV7fm1v8D-FQ02K4pT0wkU/edit?ts=5c9012f9#

Tested:
  make test-e2e-all

Bug: 128708115
Change-Id: I38099e5964cbe12b9dbf3ff0b8da782657efe39c