Consolidate functors
We have a *lot* of duplicate code which modifies objects in the same
ways: adding/removing annotations/labels, setting fields, and so on.

This CL consolidates the FileObject functors so that they can be used by
non-test code, and so that the ones written for importer and the ones
written for testing are the same. Notably, fake.BuildOpt and
mutate.Mutator serve identical function.

By opening these libraries to the entire repo, we won't have to
constantly do things like check to see if the annotations map is nil
before writing to it, or remember to manually copy maps when assigning
labels.

Change-Id: Ia0c647c46a3a9684ac73ac3ec1e946d541ed74f6