sync namespaces in a repo w/ only namespaces
This change handles a corner case where a user creates an empty repo and
starts adding only namespaces to it. Because we wait for a Sync to be
created before creating any NamespaceConfig or ClusterConfig
controllers, none of these namespaces will get synced.

The fix is to have the Sync meta controller watch for NamespaceConfig
changes and start up the Namespaceconfig and ClusterConfig controllers
if none have been started already.

TESTING: Ensured e2e tests passed, added a test for this specific case.

Fixes: 129019553
Change-Id: I4e01d5e674058e54b235111dd54a3119e48a3346