Include namespace in error messages
If a resource has a namespace, print it to the console.

This CL does not modify the logic around assigning metadata.namespace to
resources, only how/whether we print it. With current practice, this
means in general resources will only have a namespace line if they
define it in the repo or it is a namespaced resource on the cluster.

Change-Id: Iaab9137892edb589b5a22d404a35ab09ca452d00