e2e: do not delete 'default' and 'kube-system'
The nice utility for deleting resources in our e2e setup did not know
that 'default' and 'kube-system' namespaces can not be deleted.

Having those namespaces in a delete command makes the entire command
fail.  This change takes the special namespace transparently out of the
attempted delete so other namespaces could be dealt with.

It seems to me as a reasonable way to stop worrying about the special case.

This change removes those namespaces

Bug: 129299067
Change-Id: Id7215641c00d813f3e41e905a6fb6b676b78a719