Refactor applier logic
Consolidate identical logic for NamespaceConfig and ClusterConfig.
There's a lot that's essentially repeated that doesn't need to be.

This refactoring effort will reduce the scope of changes to management
state behavior and ensure we don't have to track nearly-parallel code.

This CL makes the handleDiff method independent of whether it is
applying namespace or cluster objects, and makes it not require context.

Related, ClientApplier only uses the DynamicClient for CUD operations
rather than mixing Client and DynamicClient.

E2E tests pass.

Bug: 129358726

Change-Id: I62026d5e40a6566376277b2f557b43fec6f55f7f