add error checking for unsyncable resources
Adds an error and a validator checking for abstract namespaces that have
resources, have a namespace directory as a descendant.

I had to fix some TestParser tests to pass, which were relying on the
old behavior of having unsyncable resources in abstract namespaces. For
some other cases that explicitly put unsyncable resources in an abstract
namespace directory, I had them return an error.

Fixes: 128844174
Change-Id: I54e6473dbd7145cbbc7b07b124e66f039f734050