e2e: fix kube-system and default harder
Turns out my method for not deleting resources wasn't good enough.
Apparently when you remove an item from a bash array, its length does
not change.  This leads sometimes to a nonempty array of all blanks,
which of course makes for an invalid `kubectl delete` command.

What I think this change does is (1) adds the short name for namespaces
which I missed the first time around (2) re-parses the $names[@]
array to evict elements which are all whitespace or empty, (3)
initializes "names" array with a wrong type to prevent names=() being
characterized as an unbound variable.

Bug: 128020009
Change-Id: Iad1d30c98388891c2eb5c2e0ee3332f2bdbc5d3b