syncer: ignore objects with owner references
This is meant to handle the case where a user syncs an object that
leads to the creation of other objects ownerRef'd to the synced object.
The child objects will inherit the parent's annotations, including the
nomos managed annotation.

We need to prevent attempting to delete these child objects when
reconciling.
We still need to allow updates. For example, someone could sync a
deployment and a corresponding replica set, where the replica set does
not have owner references set. The Deployment Controller will set the
owner references for the synced replica set to point to the deployment.
We still need to enforce any fields being set on the replica set in the
repo (e.g. labels).

Fixes: 129493675
Change-Id: Iea8f0c899ec9e0ee80755b8ba1ac8a13d455b92a