better error handling for invalid clusterconfig name
Before the code would retry reconciling a deleted ClusterConfig with an
invalid name. This issue would come up when upgrading Nomos after we
changed the name of the ClusterConfig.

This makes it so the error message is a little more informative and we
no longer retry reconciling deleted ClusterConfigs with invalid names.

Fixes: 129555724
Change-Id: If8fc9fe5d8716a44e10e94ce47aace277d0acfb0