e2e-stresstest: adds a single stress test for Nomos
This first test required extending the e2e test a bit to support a new
test shard.

Two new targets are added:

- `make ci-stresstest-e2e`: intended to be ran from a CI environment.
- `make ci-stresstest-e2e-local`: intended to be ran locally.

The first test is a bit simplistic: it commits 1000 namespaces and
commits them to the repo, then waits until they are all created.

Design proposal is here:
go/design-proposal-nomos-stress-and-longevity-testing

Bug: 128020009
Change-Id: I677dc3de86e64a72d567e1ffcbdf7611a4955085