Unify labeling and annotating across library
We have a lot of code that reimplements the logic of setting labels and
annotations. Everywhere that sets a label has to account for that the
label map might not exist.

This CL extracts labeling and annotating objects into the object library,
removing many instances of duplicated logic around labels and annotations.

This CL also moves specific annotation/labeling logic to the place it is
actually used. In many cases, a method dealing with a particular
annotation is used in exactly one place, but is declared in a far-off
part of the codebase.

Change-Id: I37a50a551047ad1bf2d66b2db1ac25070d654b4d