importer: allow a reconcile on sync delete
Allows an importer poll to proceed if there is a mismatch between the
syncs that we know we have and the syncs that are available on the
cluster.  We let the Sync removal succeed before we attempt to recreate
it.

Old code would only trigger on directory changes, which would miss
user-initiated sync removals on the cluster if the repo directory did
not change as well.

I interleaved some changes to the test fixture and debug logging that
made insight into the operation of the controller a bit clearer for an
unfamiliar observer.

Tested:
  make test
  make test-e2e-all

Bug: 120288321
Change-Id: Id6db33f116c7d6bca280c6b7ccc215fc89a8cc51