e2e: deflakes the "RoleBindings enforced" test
For some reason, the instantaneous check for enforced role bindings
is flaky.  It's flaky on the e2e test fixture, but even more
in the presubmits.

The root may be that the role bindings are enforced with some (small)
delay, which is not long enough on e2e presubmits.

I converted the tests to keep retrying until the assertion becomes
true.

Tested:
  make test-e2e-dev-git \
    E2E_FLAGS="--preclean --clean --setup --test_filter=enforced --file_filter=basic"

Bug: 129894255
Change-Id: Ide7236c3b7abea22d3d887ef6865e62b691f7ca4