Handle a race condition when updating latest sync token.
The syncer calculates its latest token based upon the current import
token and all configs that are currently unreconciled. It is possible
for the importer to update the import token after the syncer starts
calculating latestToken and before it writes it out. The built-in retry
for updating status will write the out-dated sync status along with the
updated import token. On the next syncer run, we can detect this
condition by checking if the new import token was seen in the previous
set of unreconciled configs.

Bug: 129062018
Change-Id: Ife69954627fd34bc0dde50eb95a7fcef1ad4ae93