Add the sync token annotation to NamespaceConfigs
We don't populate NamespaceConfigs with the SyncToken annotation, so the
Namespaces instantiated from them do not include the SyncToken
annotation.

This CL puts assigning the sync token annotation into namespaceconfig.go,
making it impossible to create or update a Namespace without using the
the SyncToken of NamespaceConfigs.

E2E Tests Pass.

Bug: 130185302
Change-Id: I5429a44cb841a9b6585ede5fa723e0c848f6ac81