syncer: update last-applied annotation on updates
During the change to ommit the nomos managed label in the last-applied
annotation, we erroneously stopped updating the last-applied annotation
on updates.

What happened was the modified version of the object used in
the 3 way merge no longer had its current state set in the last-applied
annotation. This made it so, it was never carrying forward its changes
in the last-applied annotation when the resource was modified.

Added tests to check for field, annotation, and label changes

Fixes: 130241236
Change-Id: Ief00bbcddb1b9caa6e7db791d694a8413a304d5a