git-sync: make git-sync log subprocess commands
This makes debugging somewhat easier.  There is a PR in flight
to set the log level for this info to 3, but it takes a while
to get approved.  This gives a similar effect at the cost of
some additonal log noise.

I think it should work well for our purpose.

This change needs to be released in the nomos-operator too.

Bug: 130054605
Change-Id: Ia14fbaed85244eae7f37ae7dfb011d259a343c84