importer: forbid ConfigManagement objects
Most places already don't allow ConfigManagement, the only
spot that still does is the namespaces/ dir, so added there.

In the process, found that we need the ability to create checks and
fakes for types that we don't have access to via go typing, so added
the necessary infrastructure. Via unstructured.Unstructured.

Added test for namespaces_kind_validator, which didn't have the
appropriate test.

ConfigManagement objects can still live in say the top level directory,
but that's OK.  We don't want to be copying the objects into the
cluster.

Bug: 130251683
Change-Id: I3b8a4ba0ae1988cffa71768f60f3858dacb04c56