importer: disallow deleting CRD and not CRs
Move APIInfo generation into visitor.
Create a visitor for storing CRDs pending delete.
Actually write ClusterPolicy CRD out, if it exists.

If the importer allows a commit to go through that removes a CRD, but
does not remove the corresponding CRs, it will essentially crash Syncer.
Syncer will try to add watches in the NamespaceConfig and ClusterConfig
controllers for resources that no longer exist, making it so those
controllers never start successfully.

This makes it so we look at the current CRD ClusterConfig and see if any
CRDs are going to be removed in the latest commit. We check to see if
any Custom Resources corresponding to that CRD are still in the repo, and
fail in that case.

Fixes: 129151993
Change-Id: I1ddd3ae0035d7a565ccb8bc3bd5817cf100f68eb