Remove panics from Extension
Define ast.Add and ast.Get on Extension, which return errors rather than
panicking.

This CL simply removes all unused features of Extension. Put intent of
Copy() in Add as it no longer needs to be its own separate function.

Other than for the now-depreacted Get and Set, this CL removes all
panics from Extension. The new Get returns an error if the value is not
present in the Extension. The new Set returns an error if the value is
already present in the Extension.

This CL also transitions all Nomos usages of Add and Get to the new
methods, leaving the old ones in place to allow for transitioning.

Bug: 122740271
Change-Id: Ic3acd0f542705b486fae1893ebbe5ab800de2a8b