Validate Nomos CRDs are parsed correctly
The Kubernetes object reader is happy to parse invalid Repos & etc. into
Unstructureds without throwing an error. Specifically, it is happy to do
so with any non-validated CRDs.

This CL adds ParseValidator which returns ObjectParseErrors when it
encounters configs that should have been parsed into objects rather than
Unstructureds.

ParseValidator runs before other visitors and is a requirement to
continue. Otherwise we do things like report that the Repo is missing
when really it just wasn't parseable into the Go type.

Change-Id: Icf2cb1a0c9fa344e973a5552efa8694f4f0efca5