Update RepoStatus controller to only look for unreconciled commits.
Previously the controller would try to make an educated guess about what
the "latest" sync token might be despite the presence of unreconciled
commits. This is complicated, not always correct, and inadequate to
handle NamespaceConfig deletions. Now it just looks for unreconciled
commits and if there are none, then it updates the sync token to be
equal to the import token.

Bug: 118815829
Change-Id: I8b40dc178dc713f5ccd274745a6e1bd4db6230e3