Serialize errors into repo status.
This exposes the error codes, source paths or resource details to
the user through "status" fields.  Implementation notes here:
go/structured-errors-in-nomos-implementation-notes

It is a giant change but it is by and large mechanical.

Summary of changes:

- Changed ConfigManagementError to add new fields for resources, code
  and paths.

- Changed ConfigManagementError to factor our ErrorResource from
  ConfigManagementError.  While the API is v1, the changes should not
  have changed the HTTP API surface so we should be OK to modify it. If
  not, it's gonna be OK to put it back.

- Noticed that though we have many individual errors, we have only three
  error classes.  So made three separate "error to API" converters in
  pkg/status/error.go:
    - ToErrorResource: converts a "regular" resource to an error.
    - FromError: converts from a general error
    - FromPathError: converts from a path error (notes the paths)
    - FromResourceError: converts from a resource error (fills in the
      offending resource gvk and name).

- Added a new interface ToCMEr and implemented it across all the errors.
  I think there is space to factor common code in each error classes and
  have three implementation classes, but I'd leave that for another
  commit, or leave out altogether for now.

In addition, we stopped using the field
ConfigManagementError.ErrorResource in favor of the newly introduced
ErrorResources array.  We believe this will be OK since we had no
official releases since the error fields were introduced.  The field
remains in the API so as not to break it, but is unused.

Bug: 126598308
Change-Id: Ia56578dbbc50d3304e7b71d1f84e896c7644c128