Upgrade the clientgen script to K8S 1.12.3.
This brings the clientgen script in line with the K8S version we use
in the rest of the code base.  This version should have been upgraded
when we migrated the vendor libraries. But didn't, so we're doing it
now.

The compatibility matrix is here:
https://github.com/kubernetes/client-go#compatibility-your-code---client-go

Looking at the compatibility matrix, client-go 9.0 is the client to use
for k8s 1.12 (which we do).  It is also the oldest version of
client-go that is still supported, so we really should upgrade.

However, the compatiblity guarantee at
https://github.com/kubernetes/client-go#compatibility-client-go---kubernetes-clusters
makes me think that this upgrade should not affect older client
versions.

Bug: 126598308
Change-Id: I6de8dcf497efbb956e53e96929bb3dd49b20fbc9