Swaps the return values on token_updated
The old code for examining tokens happened to pass only when the cluster
is slow to update the token such that it passes the check before the
token gets updated.

Inverting the condition and rerunning the tests, they start passing
fairly robustly.

We return zero for OK and nonzero for error.  *facepalm*.  I should
learn how return values work.

Tested:
  while true; do make test-e2e-dev-git E2E_FLAGS="--test --file_filter=status_monitoring"; done

Bug: 131339512
Change-Id: I2d1596d3dc3d26bfd1fc28366423a1dfd63879b1