Populate repo.status.sync.* on errors.
The old code has a check that compares the new candidate token
and a current one.  If the tokens are equal, then the status is
not updated.

However, this code path assumes that the status updates are atomic,
which they are not and don't need to be.  This in turn means we have to
recompute the new status every time there is a change of the watched
resources so that we can "edge trigger" properly.  A small optimization
is that we don't need to write the updated status back if it didn't
change, so added that code into the repo client.

This change should not change the performance much, as most of the
content is updated from cache anyways.  It will result in the syncer
spending more CPU.

A separate issue of too many updates remains, and this one proved a bit
tricky to figure out.  However, while we should definitely fix it so
that we don't trigger unneeded updates, we should not allow the syncer
to propagate wrong status.

Tested:
  make test-e2e-all

Bugs: 131250908
Change-Id: I139bbbbd80c700a92d017cd30a405d4268f445e7