importer trys all actions, collects errors
Instead of stopping the processing of actions when the first erroring
action is encountered, this change tweaks the importer so that it will
attempt to process every action, and return the set of errors on
failure. All the errors will then be printed to the importer log.

The impetus for this change is namespace/cluster-config finalization
work, e.g. https://b.corp.google.com/issues/130737137

Change-Id: I66949c2c9dbb1b635930b3b70ee60d0709460f3e