importer: validate CRs using in-repo CRDs as fallback
When looking up information for Custom Resources in the repo, consider
the CRDs that are currently in the repo. This is necessary because a
user could add a CRD and the corresponding CRs in the same commit.

We parse the cluster/ directory as a first pass in Parser and retrieve
all the CRDs in the repo. We update the APIInfo and RESTMappings with
the resource information in the CRDs. This requires us to stub out the
ClientGetter and RESTMapper with a version that falls back to looking at
in-repo CRD information.

Also includes a bug fix in Syncer where the ClusterPolicy controller was
reconciling CRDs.

Fixes: 129150859
Change-Id: I2effece1c92d7925d094c7a8ab2a9c33b1551c32