Define resource_error_impl to eliminate boilerplate
Also refactor last PathErrors over to the new library. Deprecate using
status.Format for PathErrors so future writers don't use the now-old
library. (The panic wouldn't be necessary if we could properly use the
deprecate annotation)

All that should be left is fairly mechanical changes of moving to the
new library.

Change-Id: I5828ea6a960c641b4c7ffa442679f64aa92b2269