importer: handle cluster CR and CRD in same commit
When generating the list of CRDs in the repo, we would fail if a Custom
Resource is in the cluster/ folder and the not yet synced CRD is also in
that folder. This is because the Custom Resource needs information from
the CRD, that we haven't yet parsed. It will fail to load the
restmapping for the Custom Resource.

The fix is to stub a restmapping for Custom Resources when are getting
the list of CRDs in the repo.

Bug: 129152433
Change-Id: I58cf593244f04f13cd6836707786a2d3e46b2699