Attempt to fix kind/CI environment
kind broke us due to https://github.com/kubernetes-sigs/kind/issues/509
but you'll node that in this Dockerfile, filip attempted to use a
specific release and thus isn't following the unrecommended path of
pinning to master. However, we can't even go get here because it tries
to install, which breaks. Hopefully if we just -d (download only) this
will be enough, and we run go install later anyway on the version we
want.

Change-Id: Id35201efeb08fdfcf2adf61a3e87fe35596d01a0