importer: disallow invalid crd names
There was a corner case where if you added CRs and a corresponding CRD
in a single commit, with an invalid name, you wouldn't be able to correct
the issue by fixing the name in the subsequent commit. This is because
of the fact that we disallow removing a CRD if the corresponding CRs are
still in the repo. Since CRDs are uniquely identified by name, we have
to assume a CRD was removed when the rename fix commit goes in.

This makes it so importer will fail validation when creating a CRD with
an invalid name, preventing a user from stepping into this mess.

Fixes: 129151993
Change-Id: Ic19b92a460c5cc25cdfdfd955f17e5f3a1ff3919