Standardize writing tabular output
The tabular output in our CLI is inconsistent and needs to be
standardized.

This CL uses text/tabwriter as the standard for writing tabular output
to the screen.

Fix bug in nomos status for when there is no token to display.
Previously it would result in a slice out of bounds and panic.

Change-Id: I13ce0710573356eac5df67855ceb6cec05505c0c