e2e: make preserve_fields tests less flaky
These tests sync all of the acme example, even though it isn't
necessary. It seems like this could be resulting in timeouts because it
sometimes takes more than 30 seconds to sync the repo and the configmaps
we add for testcases.

We also set the clustername specifically in a testcase that depended on it
being set by previous test cases.

Bug: 131429115
Change-Id: Iafc1c6558e50523bc20509299dfca4745551911c