Update nsconfig spec.Token when deletion happens
We were not updating this because the outputvisitor was not running on
the deleted node, and thus was not updating the token. So, we need to do
this token update out of band of parsing. The issue with this lack of
update is that the RepoController erronously reported that everything
was synced to the new commit even though a delete was in progress
because the spec.token and status.token matched on the nsconfig for the
ns getting deleted.

Bug: 132888992
Change-Id: Ia6bf4e8b62ba43f148e0d5e9a2e6ba6114aced1f