Update AllKubectlConfigs() to return partial results.
Previously it would fail fast if any entry in the kubeconfig file was
invalid or failed to generate a rest config. Now the function returns
all valid configs in addition to an error documenting the invalid ones.
Calling functions can decide if they should treat partial failure as
total failure or not.

Bug: 132788583
Change-Id: Iecf529f139951523b56232e44cea7de91908b7bc